+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## 18.4.2 - 2026-03-10
+* [#1811](https://github.com/stripe/stripe-ruby/pull/1811) Retry on Net::HTTPFatalError
+  - Fixes a bug where `Net::HTTPFatalError` errors were being thrown and not retried (unlike other 5xx responses from the Stripe API)
+  
+
 ## 18.4.1 - 2026-03-06
 * [#1805](https://github.com/stripe/stripe-ruby/pull/1805) Add Stripe-Request-Trigger header
 * [#1803](https://github.com/stripe/stripe-ruby/pull/1803) Add agent information to UserAgent